Split grammar/parser into a separate file
